fix(arch): move projectDefaultFilesMapSchema out of types/IPC/
`backend/shared/project/create-project-files.ts` lives in the
cross-repo shared tree (byte-identical between openplc-editor and
openplc-web), but it imported `projectDefaultFilesMapSchema` from
`@root/types/IPC/project-service` — an Electron-only namespace that
doesn't exist on web.  That made the openplc-web typecheck fail the
moment shared was synced from rc1.

The schema itself never depended on IPC: it's a plain Zod-schema map
that describes the on-disk shape of a project, used by both the
shared `createProjectFiles` (which builds the in-memory payload) and
the Electron-side project service (which writes / reads files).
Moved it to `backend/shared/project/project-files-schema.ts` next to
its main consumer and updated the deep import in the editor service.

`types/IPC/project-service/index.ts` re-exports the schema from the
new shared location so the existing IPC-namespaced imports across
the Electron-only code keep resolving — no churn for callers.
